### 2022-03-18 3a5e180
Modified and fixed normalize() for EchelleSpectrumList so that calling it will correctly normalize the spectrum to a user provided order.  Also updated normalize in EchelleSpectrum so that the value from EchelleSpectrumList can be passed to individual orders.  Previously EchelleSpectrumList did a simple divide.
