### 2022-03-18 989149a
Expand the Continuous Integration build matrix to include more modern specutils and astropy
