+++ .github/workflows/muler-tests.yml
@@ -10,8 +10,8 @@ jobs:
       fail-fast: false
       matrix:
         python-version: [3.7, 3.8]
-        specutils-version: [1.2, 1.4.1]
-        astropy-version: [4.1, 4.3]
+        specutils-version: [1.3.1, 1.4.1, 1.5, 1.6, 1.7]
+        astropy-version: [4.1, 4.3, 5.0]
         # numpy-version: [1.21.4]
         include:
           - python-version: "3.6"
