### 2022-03-18 9f4c124
Added method flatten to EchelleSpectrumList
