### 2022-03-18 bc743cc
Added append() functionality to EchelleSpectrumList to combine two spectra.  For example, combine the H and K band from IGRINS into one object.
