### 2022-03-18 82fc3a7
Removed append() from EchelleSpectrumList.  Going to implement as a utility function instead.
