### 2022-03-18 3006edb
Created definition in ultilities.py called combine_lists().  Combines two EchelleSpectrumList objects.  Useful for combining the IGRINS H and K bands.
