### 2022-03-19 8767b8c
Removed resample_list() from eschelle.py.  Use what is in utilities.py instead.
