### 2022-03-21 632f645
Reword the docstring and comments
