### 2022-03-21 3db020c
Add unit tests and some tweaks to the PR-- see GitHub conversation
