+++ src/muler/utilities.py
@@ -198,20 +198,19 @@ def resample_list(spec_to_resample, specList, **kwargs):
         spec_out[i] = spec_to_resample.resample(specList[i], **kwargs)
     return spec_out
 
-def combine_lists(EchelleSpectrumListObj1, EchelleSpectrumListObj2):
+
+def combine_lists(spec_list1, spec_list2):
     """
     Combine two EchelleSpectrumList objects into one.
     For example, combine IGRINS H and K bands.
 
     Parameters
-    EchelleSpectrumListObj1: EchelleSpectrumList object
-        Echelle spectrum with multiple orders in which to append the orders
-        from another object onto.
-    EchelleSpectrumListObj2: EchelleSpectrumList object
-        Echelle spectrum with multiple orders to append to EchelleSpectrumListObj1.
-    __________
+    ----------
+    spec_list1: EchelleSpectrumList object
+        Echelle spectrum with multiple orders
+    spec_list2: EchelleSpectrumList object
+        Echelle spectrum with multiple orders to append onto the first list.
     """
-    combinedEchelleSpectrumListObj = copy.deepcopy(EchelleSpectrumListObj1)
-    for i in range(len(EchelleSpectrumListObj2)):
-        combinedEchelleSpectrumListObj.append(EchelleSpectrumListObj2[i])
-    return combinedEchelleSpectrumListObj
+    combined_list = copy.deepcopy(spec_list1)
+    combined_list.extend(spec_list2)
+    return combined_list
