### 2022-03-21 41056dd
Changed argument median_flux to normalize_by.
