### 2022-03-21 ecce467
Updated test_utilities.py to change the name of combine_lists to concatenate_orders.
