### 2022-03-22 7629073
Fixed bug where the signal-to-noise ratio and uncertainty were not being correctly read in for IGRINS spectra.
