### 2022-03-22 f0a5104
Fix bug where doing band math does not pass 'x_values' keyword in meta for EchelleSpectrum objects.  Set plot yrange to have ylo=0 and automatically find yhi based on spectrum's flux.
