### 2022-03-22 faf8601
Minor corrections to echelle.py
