### 2022-03-22 5d56324
Minor fix to order of arguments for np.nanpercentile() in echelle.py
