### 2022-03-22 16db9be
Tweaked percentile value to for automatically finding yhi in plots.
