### 2022-03-25 345e73f
Adding method to sort stitched spectra by wavelength
