### 2022-03-25 3e2ac7b
Add a unit test, proves that the sorting method works
