### 2022-03-25 abde93c
The special x_values metadata needs to get sorted too
