### 2022-03-25 0ff9e15
Methods for EchelleSpectrumList now act on and return a copy of the spectrum instead of acting on itself.
