+++ src/muler/echelle.py
@@ -645,22 +645,21 @@ class EchelleSpectrumList(SpectrumList):
 
     def normalize(self, order_index=0):
         """Normalize all orders to one of the other orders"""
-        index = self.normalization_order_index
-        median_flux = copy.deepcopy(np.nanmedian(self[index].flux))
+        spec_out = copy.deepcopy(self)
+        index = spec_out.normalization_order_index
+        median_flux = copy.deepcopy(np.nanmedian(spec_out[index].flux))
         for i in range(len(self)):
-            self[i] = self[i].divide(median_flux, handle_meta="first_found")
+            spec_out[i] = spec_out[i].divide(median_flux, handle_meta="first_found")
 
-        return self
+        return spec_out
 
     def remove_nans(self):
         """Remove all the NaNs"""
-        # TODO: is this in-place overriding of self allowed?
-        # May have unintended consequences?
-        # Consider making a copy instead...
-        for i in range(len(self)):
-            self[i] = self[i].remove_nans()
+        spec_out = copy.deepcopy(self)
+        for i in range(len(spec_out)):
+            spec_out[i] = spec_out[i].remove_nans()
 
-        return self
+        return spec_out
 
     def remove_outliers(self, threshold=5):
         """Remove all the outliers
@@ -670,17 +669,19 @@ class EchelleSpectrumList(SpectrumList):
         threshold : float
             The sigma-clipping threshold (in units of sigma)
         """
-        for i in range(len(self)):
-            self[i] = self[i].remove_outliers(threshold=threshold)
+        spec_out = copy.deepcopy(self)
+        for i in range(len(spec_out)):
+            spec_out[i] = spec_out[i].remove_outliers(threshold=threshold)
 
-        return self
+        return spec_out
 
     def trim_edges(self, limits=None):
         """Trim all the edges"""
-        for i in range(len(self)):
-            self[i] = self[i].trim_edges(limits)
+        spec_out = copy.deepcopy(self)
+        for i in range(len(spec_out)):
+            spec_out[i] = spec_out[i].trim_edges(limits)
 
-        return self
+        return spec_out
 
     def deblaze(self, method="spline"):
         """Remove blaze function from all orders by interpolating a spline function
@@ -689,15 +690,15 @@ class EchelleSpectrumList(SpectrumList):
                 otherwise  effects can be appear from zero-padded edges.
         """
         spec_out = copy.deepcopy(self)
-        for i in range(len(self)):
-            spec_out[i] = self[i].deblaze(method=method)
+        for i in range(len(spec_out)):
+            spec_out[i] = spec_out[i].deblaze(method=method)
         return spec_out
 
     def flatten_by_black_body(self, Teff):
         """Flatten by black body"""
         spec_out = copy.deepcopy(self)
-        index = self.normalization_order_index
-        median_wl = copy.deepcopy(np.nanmedian(self[index].wavelength))
+        index = spec_out.normalization_order_index
+        median_wl = copy.deepcopy(np.nanmedian(spec_out[index].wavelength))
 
         blackbody_func = BlackBody(temperature=Teff * u.K)
         blackbody_ref = blackbody_func(median_wl)
