### 2022-03-30 #4220 @xuanxu
@eteq / @saogaz / @brechmos: would you be willing to review this paper? As specutils contributors all of you are a great fit for this submission.
