### 2022-03-30 02c140d
Modified reading in IGRINS uncertainity by reading from .variance.fits instead of .sn.fits files, because the sn.fits files give the signal-to-noise per resolution element, and not pixel.
