### 2022-04-13 40f1204
Create CODE_OF_CONDUCT.md
