### 2022-04-22 b182249
Some tweaks to the normalize code, some tests are failing and we should add more
