### 2022-04-25 8b1e1bd
Tweaks to the source code and tests, non-standard copying/in-place adjustments...
