### 2022-04-25 ec817c2
add a bunch of new tests for normalization, still slightly non-intuitive for spectrumlist-like objects
