### 2022-04-25 1c1f893
Minor docstring edits, PR should be good to go
