### 2022-04-25 ddb975d
Updates to normalize() (#81)
* Modified and fixed normalize() for EchelleSpectrumList so that calling it will correctly normalize the spectrum to a user provided order.  Also updated normalize in EchelleSpectrum so that the value from EchelleSpectrumList can be passed to individual orders.  Previously EchelleSpectrumList did a simple divide.

* Added method flatten to EchelleSpectrumList

* Added append() functionality to EchelleSpectrumList to combine two spectra.  For example, combine the H and K band from IGRINS into one object.

* Removed append() from EchelleSpectrumList.  Going to implem
[…]
