+++ src/muler/echelle.py
@@ -27,7 +27,7 @@ from scipy.signal import savgol_filter
 from astropy.constants import R_jup, R_sun, G, M_jup, R_earth, c
 from astropy.modeling.physical_models import BlackBody
 import specutils
-from muler.utilities import apply_numpy_mask, resample_list
+from muler.utilities import apply_numpy_mask
 
 # from barycorrpy import get_BC_vel
 from astropy.coordinates import SkyCoord, EarthLocation
@@ -151,30 +151,56 @@ class EchelleSpectrum(Spectrum1D):
         ew = equivalent_width(self, regions=SpectralRegion(lower, upper))
         return ew
 
-    def normalize(self):
-        """Normalize spectrum by its median value
+    def normalize(self, normalize_by="median"):
+        """Normalize the spectrum by a scalar value, usually its median
+
+        Parameters
+        ----------
+        normalize_by : (string or float or Quantity)
+            The flux value or method name to normalize by.  Usually this is the 
+            median flux "median", making the resulting flux vector have a median 
+            value of 1 (default).  The user may optionally pass in the strings
+            "mean" for normalization by the mean flux, or "peak" for normalization
+            by the 90th percentile of the spectrum.  If a Quanitity is passed in,
+            it must have equivalent flux units as the spectrum.
 
         Returns
         -------
-        normalized_spec : (KeckNIRSPECSpectrum)
+        normalized_spec : (EchelleSpectrum)
             Normalized Spectrum
         """
         spec = self._copy(
             spectral_axis=self.wavelength.value * self.wavelength.unit, wcs=None
         )
-        median_flux = np.nanmedian(spec.flux.value)
+
+        flux_unit = spec.flux.unit
+
+        # We default to normalizing by the median flux value
+        if normalize_by == "median":
+            normalize_by = np.nanmedian(spec.flux.value)
+        elif normalize_by == "mean":
+            normalize_by = np.nanmean(spec.flux.value)
+        elif normalize_by == "peak":
+            normalize_by = np.nanpercentile(spec.flux.value, 90.0)
+        elif isinstance(normalize_by, u.Quantity):
+            flux_unit = normalize_by.unit
+            normalize_by = normalize_by.value
+            assert flux_unit.to(spec.flux.unit) is not None, "Flux units must match"
+        else:  # must be a number
+            assert normalize_by == float(
+                normalize_by
+            ), "Must be floating point eligible"
 
         # Each ancillary spectrum (e.g. sky) should also be normalized
         meta_out = copy.deepcopy(spec.meta)
         for ancillary_spectrum in self.available_ancillary_spectra:
             meta_out[ancillary_spectrum] = meta_out[ancillary_spectrum].divide(
-                median_flux * spec.flux.unit, handle_meta="ff"
+                normalize_by * flux_unit, handle_meta="ff"
             )
 
-        # spec.meta = meta_out
-        return spec.divide(
-            median_flux * spec.flux.unit, handle_meta="first_found"
-        )._copy(meta=meta_out)
+        return spec.divide(normalize_by * flux_unit, handle_meta="first_found")._copy(
+            meta=meta_out
+        )
 
     def flatten_by_black_body(self, Teff):
         """Flatten the spectrum by a scaled black body, usually after deblazing
@@ -477,7 +503,7 @@ class EchelleSpectrum(Spectrum1D):
         else:
             return smoothed_spectrum
 
-    def plot(self, ax=None, ylo=0.6, yhi=1.2, figsize=(10, 4), **kwargs):
+    def plot(self, ax=None, ylo=0.0, yhi=None, figsize=(10, 4), **kwargs):
[…]
