### 2022-04-25 57f9d08
Adding method to sort stitched spectra by wavelength (#85)
* Adding method to sort stitched spectra by wavelength

* Untested Sorting Method

* Add a unit test, proves that the sorting method works

* sorting even works for HPF out of the box, even ancillary sky fiber

* The special x_values metadata needs to get sorted too

* Half-baked paper about data reduction, work in progress

Co-authored-by: gully <igully@gmail.com>
