### 2022-04-25 6fd174f
Methods for EchelleSpectrumList now act on and return a copy of the spectrum instead of acting on itself. (#87)
Co-authored-by: gully <igully@gmail.com>
