### 2022-04-25 f48e214
use is close not exactly equal for compairing floats
