### 2022-04-25 8597382
Astropy 5 and python 3.7 still do not mix, how to prevent this combo?
