### 2022-04-25 a38bbdc
Experiment to see what RTD will do if I ask it to render notebooks, Issue #93
