### 2022-04-25 d286fd4
nbsphinx needs to know about ipython kernels
