### 2022-04-26 571ee6b
The sort method was missing uncertainty propagation for ancillary spectra
