+++ src/muler/echelle.py
@@ -840,9 +840,23 @@ class EchelleSpectrumList(SpectrumList):
             fluxes_anc = np.hstack(
                 [spec[i].meta[ancillary_spectrum].flux for i in range(len(spec))]
             )
+            if spec[0].meta[ancillary_spectrum].uncertainty is not None:
+                # HACK We assume if one order has it, they all do, and that it's StdDev
+                unc_anc = np.hstack(
+                    [
+                        spec[i].meta[ancillary_spectrum].uncertainty.array
+                        for i in range(len(self))
+                    ]
+                )
+                unc_anc = StdDevUncertainty(unc_anc)
+            else:
+                unc_anc = None
 
             meta_out[ancillary_spectrum] = spec[0].__class__(
-                spectral_axis=wls_anc, flux=fluxes_anc, meta=meta_of_meta
+                spectral_axis=wls_anc,
+                flux=fluxes_anc,
+                uncertainty=unc_anc,
+                meta=meta_of_meta,
             )
 
         return spec[0].__class__(
