### 2022-04-26 dc8f824
Experiment to get RTD to fetch the example data into the right place...
