+++ .readthedocs.yaml
@@ -1,0 +1,14 @@
+version: 2
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.8"
+  jobs:
+    pre_create_environment:
+      - echo "Command run at 'pre_create_environment' step"
+      - pwd
+      - which git
+      - ls
+      - git clone https://github.com/OttoStruve/muler_example_data.git
+      - ls
