### 2022-04-26 fbe48e1
attempt some advanced bash to get example data to work
