+++ .readthedocs.yaml
@@ -8,10 +8,11 @@ build:
     pre_create_environment:
       - echo "Command run at 'pre_create_environment' step"
       - pwd
+      - cwd=$(pwd)
       - ls
       - cd ..
       - git clone https://github.com/OttoStruve/muler_example_data.git
-      - cd muler
+      - cd cwd
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
