### 2022-04-26 2a8093a
bash patterns...
