+++ .readthedocs.yaml
@@ -8,11 +8,11 @@ build:
     pre_create_environment:
       - echo "Command run at 'pre_create_environment' step"
       - pwd
-      - cwd=$(pwd)
+      - mycwd=$(pwd)
       - ls
       - cd ..
       - git clone https://github.com/OttoStruve/muler_example_data.git
-      - cd cwd
+      - cd $mycwd
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
