### 2022-04-26 f023c4e
Standardize paths to example data in the tutorials, and (for now) always re-render notebooks with nbsphinx (#98)
* The tutorials can be forced to render online with one simple change to the conf.py docs file.

* Spot-check path inconsistencies in notebooks

* Standardize the path to data for all notebooks

* Provide instructions for symlink-ing the test data
