+++ docs/conf.py
@@ -51,7 +51,7 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**/.ipynb_checkpoints"]
 nbsphinx_timeout = 60
 
 # Execute notebooks? Possible values: 'always', 'never', 'auto' (default)
-nbsphinx_execute = "auto"
+nbsphinx_execute = "always"
 
 autosummary_generate = True
 html_show_sourcelink = True
