+++ .readthedocs.yaml
@@ -24,5 +24,10 @@ sphinx:
 
 # Optionally declare the Python requirements required to build your docs
 python:
-   install:
-   - requirements: docs/requirements.txt
+  install:
+    - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+    - method: setuptools
+      path: .
+  system_packages: true
