+++ docs/tutorials/index.rst
@@ -32,10 +32,6 @@ HPF Examples
 
    Precision Sky Subtraction with Blank Sky Observations <blank_sky_observations.ipynb>
 
-   Masking and continuum flattening with specutils <Masking_and_flattening.ipynb>
-
-   Masking and continuum flattening with TelFit and Celeritè <Masking_and_flattening_plus.ipynb>
-
    Combining misaligned spectra <Combine_spectra_misaligned.ipynb>
 
    
@@ -62,9 +58,3 @@ Concepts
 
    Easy Flattening with built-in Savitzky-Golay filter <Flattening_with_savgol.ipynb>
 
-Deprecated
-##########
-.. toctree::
-   :maxdepth: 1
-
-   HPF Sky Subtraction and Deblazing <HPF_tutorial.ipynb>
