### 2022-04-26 0094277
Overhaul tutorials to use HTML paths, also use nbstripout (#99)
* overhaul tutorials to use HTML paths, also use nbstripout

* HTML paths across the board

* Swap to html paths for tutorials

* Remove duplicate tutorials

* Delete RTD file, it should work out-of-the-box now
