+++ .readthedocs.yaml
@@ -1,33 +1,0 @@
-version: 2
-
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.8"
-  jobs:
-    pre_create_environment:
-      - echo "Command run at 'pre_create_environment' step"
-      - pwd
-      - mycwd=$(pwd)
-      - ls
-      - cd ..
-      - git clone https://github.com/OttoStruve/muler_example_data.git
-      - cd $mycwd
-
-# Build documentation in the docs/ directory with Sphinx
-sphinx:
-   configuration: docs/conf.py
-
-# If using Sphinx, optionally build your docs in additional formats such as PDF
-# formats:
-#    - pdf
-
-# Optionally declare the Python requirements required to build your docs
-python:
-  install:
-    - requirements: docs/requirements.txt
-    - method: pip
-      path: .
-    - method: setuptools
-      path: .
-  system_packages: true
