### 2022-04-26 7e670e1
Hmm, mimick what lightkurve does to get tutorials to compile
