### 2022-04-26 0f11119
Allow errors in nbsphinx for now...
