+++ docs/conf.py
@@ -14,7 +14,7 @@ import os
 import sys
 
 sys.path.insert(0, os.path.abspath("../src/"))
-sys.path.append(os.path.join(os.path.dirname(__name__), ".."))
+# sys.path.append(os.path.join(os.path.dirname(__name__), ".."))
 
 import muler
 
