+++ docs/requirements.txt
@@ -6,10 +6,12 @@ sphinx-material
 scipy
 pandas
 nbsphinx
-gwcs<0.17
+gwcs
 celerite2
 matplotlib
 requests
 ipykernel
 importlib_resources
 h5py
+celerite2
+gollum
