### 2022-04-26 5cd3b94
RTD has a problem with importlib resources I do not know why
