+++ .readthedocs.yaml
@@ -27,3 +27,6 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
+   - method: setuptools
+     path: ./
+   system_packages: true
