### 2022-04-26 73e3215
hail mary, install the pip version of muler
