+++ docs/requirements.txt
@@ -15,3 +15,4 @@ importlib_resources
 h5py
 celerite2
 gollum
+muler
