+++ .readthedocs.yaml
@@ -25,8 +25,6 @@ sphinx:
 
 # Optionally declare the Python requirements required to build your docs
 python:
-   install:
+  install:
    - requirements: docs/requirements.txt
-   - method: setuptools
-     path: ./
    system_packages: true
