+++ README.md
@@ -1,6 +1,6 @@
 # `muler` (_μler_)
 
-### version 0.3.0
+### version 0.3.1
 
 <a href="https://muler.readthedocs.io/en/latest/"><img src="https://img.shields.io/badge/Read-the%20docs-blue"></a>
 <a href="https://pypi.org/project/muler/"><img src="https://img.shields.io/badge/pip_install-muler-9b59b6"></a>
