+++ docs/requirements.txt
@@ -17,3 +17,4 @@ celerite2
 gollum
 tqdm
 muler
+numpydoc
