+++ docs/requirements.txt
@@ -18,3 +18,4 @@ gollum
 tqdm
 muler
 numpydoc
+bokeh
