### 2022-04-27 acbad4a
A prototype for fix for Issue #78
