### 2022-04-27 282aba9
We have a Tutorial Gallery now! Closes #78
