+++ docs/tutorials/index.rst
@@ -3,67 +3,44 @@
 Tutorials
 ---------
 
+Beginner
+########
+
 .. nbgallery::
     :caption: This is a thumbnail gallery:
     :name: fungallery
 
     exploratory.ipynb
+    All_spectral_orders.ipynb
+    Deblazing_HPF_spectra.ipynb
     refined_sky_subtraction.ipynb
+    
+    
 
-
-
-IGRINS Examples
-###############
-
-.. toctree::
-   :maxdepth: 1
-   
-   IGRINS Hello World <exploratory.ipynb>
-
-   Manipulate all spectral orders at once <All_spectral_orders.ipynb>
-
-   Download IGRINS data from Google Drive <Download_IGRINS_data_from_GoogleDrive.ipynb>
-   
-   Workarounds for manipulating multiple IGRINS orders (intermediate usage) <IGRINS_SpecList_demo.ipynb>
-   
-
-HPF Examples
+Intermediate
 ############
 
-.. toctree::
-   :maxdepth: 1
+.. nbgallery::
+    :caption: This is a thumbnail gallery:
+    :name: fungallery2
 
-   Deblazing HPF data with an automatic standard template <Deblazing_HPF_spectra.ipynb>
+    Combining_uncertainties_with_specutils.ipynb
+    Combine_spectra_misaligned.ipynb
+    Masking_and_flattening.ipynb
+    Flattening_with_savgol.ipynb
+    
 
-   Refined Sky Subtraction <refined_sky_subtraction.ipynb>
-
-   Telluric correction with muler and gollum <telluric_correction_with_gollum.ipynb>
-
-   Precision Sky Subtraction with Blank Sky Observations <blank_sky_observations.ipynb>
-
-   Combining misaligned spectra <Combine_spectra_misaligned.ipynb>
-
-   
-
-Concepts
+    
+Advanced
 ########
 
-*Error Propagation*
-^^^^^^^^^^^^^^^^^^^
-.. toctree::
-   :maxdepth: 1
+.. nbgallery::
+    :caption: This is a thumbnail gallery:
+    :name: fungallery3
 
-   Combining uncertainties with specutils <Combining_uncertainties_with_specutils.ipynb>
-
-
-*Continuum Flattening*
-^^^^^^^^^^^^^^^^^^^^^^
-.. toctree::
-   :maxdepth: 1
-
-   Masking and continuum flattening with specutils <Masking_and_flattening.ipynb>
-
-   Masking and continuum flattening with TelFit and Celeritè <Masking_and_flattening_plus.ipynb>
-
-   Easy Flattening with built-in Savitzky-Golay filter <Flattening_with_savgol.ipynb>
+    telluric_correction_with_gollum.ipynb
+    Masking_and_flattening_plus.ipynb
+    blank_sky_observations.ipynb
+    IGRINS_SpecList_demo.ipynb
+    Download_IGRINS_data_from_GoogleDrive.ipynb
 
