### 2022-04-28 7e32fd2
Add thumbnail logos for almost all notebook tutorials in the gallery
