### 2022-04-28 cbade4c
Add @ericasaw as coauthor to the muler JOSS paper
