### 2022-04-28 #4302 @gully
Thank you all for the reviewing and editing process.  We would like to add a coauthor who contributed code during the JOSS review process.  I have made the coauthor addition on the `paper.md` in the main branch of the repo.

Thank you all again 🙏
