### 2022-04-28 OttoStruve/muler#108 @wtgee
JOSS Review - Need local paths for PHOENIXSpectrum in Masking notebook
My first run of this notebook was missing `gollum` as it's only listed in the `docs/requirements.txt` file, which is fine. I installed with `pip install gollum` (I didn't use the full `docs/requirements.txt` because I didn't want to bother with `sphinx`, etc.).

After `pip` install I get the following exception on run:

```py
In [7]: native_template = PHOENIXSpectrum(teff=9600, logg=5.0).normalize()

AssertionError                            Traceback (most recent call last)
Input In [7], in <cell line: 1>()
----> 1 native_template = PHOENIXSpectrum(teff=9600, logg=5.0).normalize()

File ~/miniforge3/envs/muler_dev/lib/python3.8/site-packages/gollum/phoenix.py:72, in PHOENIXSpectrum.__init__(self, teff, logg, metallicity, path, wl_lo, wl_hi, *args, **kwargs)
     70 if (teff is not None) & (logg is not None):
     71     base_path = os.path.expanduser(path)
---> 72     assert os.path.exists(
     73         base_path
     74     ), "You must specify the path to local PHOENIX models"
     76     wl_filename = base_path + "/WAVE_PHOENIX-ACES-AGSS-COND-2011.fits"
     77     assert os.path.exists(
     78         wl_filename
     79     ), "You need to place the PHOENIX models in {}".format(base_path)

AssertionError: You must specify the path to local PHOENIX models
```

Looks like a hard-coded path in `gollum` here:

https://github.com/BrownDwarf/gollum/blob/main/src/gollum/phoenix.py#L66

Edit: this also prevents me from running the [Telluric correction](https://github.com/OttoStruve/muler/blob/main/docs/tutorials/telluric_correction_with_gollum.ipynb) and [Masking plus](https://github.com/OttoStruve/muler/blob/main/docs/tutorials/Masking_and_flattening_plus.ipynb) notebooks.
