### 2022-04-28 #4302 @wtgee
@gully I finished off the review and added a few minor issues as well as https://github.com/OttoStruve/muler/issues/108, which prevents me from running the notebooks from a clean install. If I manually fix that issue the two notebooks don't have any other issues.

I've checked off all the boxes for my review but that Issue above should get fixed before final approval. The other minor grammatical issues don't affect the review.
