### 2022-04-29 OttoStruve/muler#108 @gully
Thank you for this feedback, I will get these notebooks working by updating the guidance on the hard-coded paths.
