### 2022-05-04 OttoStruve/muler#108 @gully
Woohoo! I have solved this issue in a slick way--- I just released `gollum v0.2.0` which introduces a [new](https://github.com/BrownDwarf/gollum/pull/38) `download=True` option for the PHOENIX synthetic spectral models.  Now muler newcomers and cloud-based venues like readthedocs can easily fetch a PHOENIX model without having to download and store the voluminous model grid locally.  Yay!
