### 2022-05-04 744d491
Drop testing for specutils <1.5.  I suppose that means we are dropping support for specutils <1.5?
