### 2022-05-04 37fa577
Require modern specutils, since it handles radial_velocity starting with 1.5
