### 2022-05-04 37f4d81
Drop support for python 3.7 entirely
