### 2022-05-04 01f1078
Drop support for python 3.7 add tests for python 3.9
