+++ .github/workflows/muler-tests.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.7, 3.8]
+        python-version: [3.8, 3.9]
         specutils-version: [1.5, 1.6, 1.7]
         astropy-version: [4.1, 4.3, 5.0]
         # numpy-version: [1.21.4]
@@ -17,10 +17,10 @@ jobs:
           - python-version: "3.6"
             astropy-version: "4.1"
             specutils-version: "1.3.1"
-        exclude:
+        #exclude:
         # Astropy 5 drops support for python 3.7
-          - python-version: "3.7"
-            astropy-version: "5.0"
+        #  - python-version: "3.9"
+        #    astropy-version: "5.0"
 
     steps:
       - uses: actions/checkout@v2
