### 2022-05-04 4f1a37e
Require tqdm for docs to render
