### 2022-05-04 de75114
Drop support for astropy<=4.1
