+++ .github/workflows/muler-tests.yml
@@ -11,16 +11,7 @@ jobs:
       matrix:
         python-version: [3.8, 3.9]
         specutils-version: [1.5, 1.6, 1.7]
-        astropy-version: [4.1, 4.3, 5.0]
-        # numpy-version: [1.21.4]
-        include:
-          - python-version: "3.6"
-            astropy-version: "4.1"
-            specutils-version: "1.3.1"
-        #exclude:
-        # Astropy 5 drops support for python 3.7
-        #  - python-version: "3.9"
-        #    astropy-version: "5.0"
+        astropy-version: [4.3, 5.0]
 
     steps:
       - uses: actions/checkout@v2
