### 2022-05-04 7f1e8c8
gollum 0.2.1 fixes the download=True/False feature needed for notebooks to run in the Cloud
