+++ docs/requirements.txt
@@ -16,7 +16,7 @@ h5py
 celerite2
 tqdm
 bokeh # needed for gollum to install on RTD
-gollum==0.2.0
+gollum==0.2.1
 muler
 numpydoc
 sphinx-gallery
