### 2022-05-04 OttoStruve/muler#108 @gully
Just a followup here, a bug had prevented this from _actually_ working, it should be fine now though! 😄
