### 2022-05-07 #4302 @xuanxu
OK @gully, everything looks good, here are the next steps:

- Please release a new tagged version from the current main branch so it includes all the changes made during the review process
- Then archive that latest release in Zenodo or a similar service
- Check the Zenodo deposit has the correct metadata: title and authors names should match the paper; you may also add authors' ORCIDs.

Once you do that please report here the version number and archive DOI
