+++ tests/test_igrins.py
@@ -151,7 +151,7 @@ def test_uncertainty_spec_a0v():
 
     snr_vec = new_spec.flux / new_spec.uncertainty.array
     snr_med = np.nanmedian(snr_vec.value)
-    #assert np.isclose(snr_med, snr_old_med, atol=0.005)
+    assert np.isclose(snr_med, snr_old_med, atol=0.005)
 
 
 def test_uncertainty_spec_fits():
@@ -185,7 +185,7 @@ def test_uncertainty_spec_fits():
 
     snr_vec = new_spec.flux / new_spec.uncertainty.array
     snr_med = np.nanmedian(snr_vec.value)
-    #assert np.isclose(snr_med, snr_old_med, atol=0.005)
+    assert np.isclose(snr_med, snr_old_med, atol=0.005)
 
 
 def test_equivalent_width():
