### 2022-05-10 c463a86
Updated to handle wave files without a full path.
