### 2022-05-10 ec48802
On the way towards making spec_a0v.fits files import correctly
